Remove unnecessary service layer Add PlamtUML schema to understand how TCP work in my application
