+++ README.md
@@ -11,13 +11,13 @@
 
 ## Описание/Пошаговая инструкция выполнения домашнего задания:
 
-Библиотека предоставляет **динамическую** структуру дома в комнатах которого расположены устройства.
+Библиотека предоставляет динамическую структуру дома в комнатах которого расположены устройства.
 
 - Дом имеет название и содержит несколько помещений.
-- Библиотека позволяет запросить список помещений в доме, **а также добавлять и удалять помещения**.
+- Библиотека позволяет запросить список помещений в доме, а также добавлять и удалять помещения.
 - Помещение имеет уникальное название и содержит названия нескольких устройств.
 - Устройство имеет уникальное в рамках помещения имя.
-- Библиотека позволяет получать список устройств в помещении, **а также добавлять и удалять устройства**.
+- Библиотека позволяет получать список устройств в помещении, а также добавлять и удалять устройства.
 - Библиотека имеет функцию, возвращающую текстовый отчёт о состоянии дома.
 - Эта функция принимает в качестве аргумента обобщённый тип, позволяющий получить текстовую информацию
 о состоянии устройства, для включения в отчёт. Эта информация должна предоставляться
@@ -25,3 +25,10 @@
 Если устройство не найдено в источнике информации, то вместо текста о состоянии вернуть сообщение об ошибке.
 Привести пример типа, предоставляющего текстовую информацию об устройствах в доме для составления отчёта.
 
+## Дополнение !!!
+
+### TCP-клиент умной розетки, позволяющий:
+
+- включать и выключать розетку,
+- запрашивать информацию о текущем состоянии и потребляемой мощности розетки.
+- Также, для проверки нового функционала, реализовать приложение, имитирующее работу умной розетки, управляемой по TCP.
