+++ home/src/network/client.rs
@@ -1,0 +1,7 @@
+struct Client {
+
+}
+
+impl Client {
+
+}
