+++ home/src/network/mod.rs
@@ -1,0 +1,2 @@
+pub mod client;
+pub mod server;
