+++ home/src/services.rs
@@ -6,234 +6,11 @@ use crate::{
 };
 use std::slice::Iter;
 
-/// Struct to store services to manage home
-///
-/// Struct contains two services:
-/// 1. Service which manage devices store
-/// 2. Service which manage schema of the home
-pub struct ServiceDeviceManagement {
-    service_devices: ServiceDevices,
-    service_schema: ServiceSchemaDevices,
-}
-
-impl Default for ServiceDeviceManagement {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ServiceDeviceManagement {
-    /// Method create new struct
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let service = ServiceDeviceManagement::new();
-    /// ```
-    pub fn new() -> Self {
-        let service_devices = ServiceDevices::new();
-        let service_schema = ServiceSchemaDevices::new();
-        ServiceDeviceManagement {
-            service_devices,
-            service_schema,
-        }
-    }
-
-    /// Method add new room to services
-    ///
-    /// Now, service add room only to schema service
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room = Room::new("My room");
-    /// let mut service = ServiceDeviceManagement::new();
-    /// let result: Result<(), String> = service.add_room(&room);
-    ///
-    /// # assert!(result.is_ok()); // normal add is OK
-    /// # assert!(service.add_room(&room).is_err()); // duplicate add is KO
-    /// ```
-    pub fn add_room(&mut self, room: &Room) -> Result<(), String> {
-        self.service_schema.add_room(room.name())
-    }
-
-    /// Method remove room from schema
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room_name = "My room";
-    /// let room = Room::new(room_name);
-    /// let mut service = ServiceDeviceManagement::new();
-    /// let result_add: Result<(), String> = service.add_room(&room);
-    /// let result_remove: Result<(), String> = service.remove_room(room_name);
-    ///
-    /// # assert!(result_add.is_ok()); // normal add is OK
-    /// # assert!(result_remove.is_ok()); // normal remove is OK
-    /// # assert!(service.remove_room(room_name).is_err()); // remove room, that not contains, is KO
-    /// ```
-    pub fn remove_room(&mut self, room_name: &str) -> Result<(), String> {
-        self.service_schema.remove_room(room_name)
-    }
-
-    /// Method add device to specific room to services
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::devices::socket::Socket;
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room = Room::new("My room");
-    /// let mut service = ServiceDeviceManagement::new();
-    /// service.add_room(&room).unwrap();
-    ///
-    /// let device = Socket::new();
-    /// let result: Result<(), String> = service.add_device(room.name(), Box::new(device));
-    ///
-    /// # assert!(result.is_ok()); // normal add is OK
-    /// # assert!(service.add_device(room.name(), Box::new(Socket::new())).is_err()); // duplicate add is KO
-    /// ```
-    pub fn add_device(&mut self, room_name: &str, device: Box<dyn Device>) -> Result<(), String> {
-        let device_name = String::from(device.name());
-        self.service_devices.add_device(device)?;
-        self.service_schema
-            .add_device(room_name, device_name.as_str())?;
-        Ok(())
-    }
-
-    pub fn remove_device(&mut self, device_name: &str) -> Result<(), String> {
-        self.service_schema.remove_device(device_name)?;
-        self.service_devices.remove_device(device_name)
-    }
-
-    /// Method connect devices
-    /// TODO: rewrite method
-    ///
-    pub fn connect_device(
-        &mut self,
-        room_name: &str,
-        device_connects_to: &str,
-        device_connected: &str,
-    ) -> Result<(), String> {
-        self.service_schema
-            .connect_device(room_name, device_connects_to, device_connected)
-            .unwrap();
-
-        Ok(())
-    }
-
-    /// Method return device from store devices by specific name
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::devices::Device;
-    /// use crate::home::devices::socket::Socket;
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room = Room::new("My room");
-    /// let mut service = ServiceDeviceManagement::new();
-    /// service.add_room(&room).unwrap();
-    ///
-    /// let device_name = "My device";
-    /// let device = Socket::from(device_name, "Description", 1000.0);
-    /// service.add_device(room.name(), Box::new(device)).unwrap();
-    ///
-    /// let device_opt: Option<&dyn Device> = service.get_device(device_name);
-    ///
-    /// # assert!(device_opt.is_some()); // normal get device return some is OK
-    /// # assert_eq!(device_opt.unwrap().name(), device_name); // normal get device is OK
-    /// # assert!(service.get_device("Unknown device").is_none()); // unknown get device is KO
-    /// ```
-    pub fn get_device(&self, device_name: &str) -> Option<&dyn Device> {
-        self.service_devices.get_device(device_name)
-    }
-
-    /// Method return iter with list of devices names in the room
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::devices::socket::Socket;
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room_name = "My room";
-    /// let room = Room::new(room_name);
-    /// let mut service = ServiceDeviceManagement::new();
-    /// service.add_room(&room).unwrap();
-    ///
-    /// let device_name = "My device";
-    /// let device = Socket::from(device_name, "Description", 1000.0);
-    /// service.add_device(room.name(), Box::new(device)).unwrap();
-    ///
-    /// let devices = service.get_devices(room_name);
-    /// # assert!(!devices.collect::<Vec<&String>>().is_empty()); // normal get devices is OK
-    /// # assert!(service.get_devices("Unknown room").collect::<Vec<&String>>().is_empty()); // unknown room get devices is KO
-    /// ```
-    pub fn get_devices(&self, room_name: &str) -> Iter<String> {
-        self.service_schema.room_devices(room_name)
-    }
-
-    /// Method collect report and print it to stdout
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::devices::socket::Socket;
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room_name = "My room";
-    /// let room = Room::new(room_name);
-    /// let mut service = ServiceDeviceManagement::new();
-    /// service.add_room(&room).unwrap();
-    ///
-    /// let device_name = "My device";
-    /// let device = Socket::from(device_name, "Description", 1000.0);
-    /// service.add_device(room.name(), Box::new(device)).unwrap();
-    ///
-    /// service.print_report();
-    /// ```
-    pub fn print_report(&self) {
-        let report = self.service_devices.collect_data_for_report();
-        println!("Generated report about all devices:\n{}", report);
-    }
-
-    /// Method collect schema information and print it to stdout
-    ///
-    /// Example:
-    /// ```
-    /// use crate::home::devices::socket::Socket;
-    /// use crate::home::places::Room;
-    /// use crate::home::services::ServiceDeviceManagement;
-    ///
-    /// let room_name = "My room";
-    /// let room = Room::new(room_name);
-    /// let mut service = ServiceDeviceManagement::new();
-    /// service.add_room(&room).unwrap();
-    ///
-    /// let device_name = "My device";
-    /// let device = Socket::from(device_name, "Description", 1000.0);
-    /// service.add_device(room.name(), Box::new(device)).unwrap();
-    ///
-    /// service.print_schema();
-    /// ```
-    pub fn print_schema(&self) {
-        let report = self.service_schema.collect_schema();
-        println!("{}", report);
-    }
-}
-
 /// Struct to manage store of device of the home
 ///
 /// Struct contains only store with all devices
 ///
-struct ServiceDevices {
+pub struct ServiceDevices {
     store_devices: StoreDevices,
 }
 
@@ -292,7 +69,7 @@ impl ServiceDevices {
 }
 
 /// Struct to manage store of the schema of the Home
-struct ServiceSchemaDevices {
+pub struct ServiceSchemaDevices {
     store_schema: StoreDeviceLinks,
 }
